Create helloworld.html
This was originally created in "Edity" by Google- in the browser. 
Then I continued editing in SoloLearn.
The content is notes about my attempts at usage and reviews of different text editors, in browser only, not the type you download.
At the end there is a statement about where I would like to see this project go.