Add files via upload
Following the tutorials provided on _SoloLearn_, taking notes, reading diligently through comments, I am nearing the end of the course. Incorporated all HTML elements that I am least familiar with on this html page, and I realize why I am less familiar with these. The elements are not as useful when not coupled with JavaScript or an API, things I am still relatively new to.  This exercise has shown me where I should be studying if I want to grow my projects.